[CSPM] extract sysprobe client into interface (in preparation for local client)
