+++ pkg/compliance/compliance.go
@@ -9,10 +9,7 @@
 package compliance
 
 import (
-	"context"
 	"fmt"
-	"net"
-	"net/http"
 	"os"
 	"time"
 
@@ -70,9 +67,9 @@ func StartCompliance(log log.Component,
 		resolverOptions.StatsdClient = statsdClient
 	}
 
-	var sysProbeClient *http.Client
+	var sysProbeClient SysProbeClient
 	if config := sysprobeconfig.SysProbeObject(); config != nil && config.SocketAddress != "" {
-		sysProbeClient = newSysProbeClient(config.SocketAddress)
+		sysProbeClient = NewRemoteSysProbeClient(config.SocketAddress)
 	}
 
 	enabledConfigurationsExporters := []ConfigurationExporter{
@@ -124,20 +121,4 @@ func sendRunningMetrics(statsdClient ddgostatsd.ClientInterface, moduleName stri
 	}()
 
 	return heartbeat
-}
-
-func newSysProbeClient(address string) *http.Client {
-	return &http.Client{
-		Timeout: 10 * time.Second,
-		Transport: &http.Transport{
-			MaxIdleConns:    2,
-			IdleConnTimeout: 30 * time.Second,
-			DialContext: func(_ context.Context, _, _ string) (net.Conn, error) {
-				return net.Dial("unix", address)
-			},
-			TLSHandshakeTimeout:   1 * time.Second,
-			ResponseHeaderTimeout: 5 * time.Second,
-			ExpectContinueTimeout: 50 * time.Millisecond,
-		},
-	}
 }
